// src/certificate/certificate.enums.ts
export enum AchievementLevel {
  BRONZE = 'bronze',
  SILVER = 'silver',
  GOLD = 'gold',
  PLATINUM = 'platinum',
}

export enum CertificateStatus {
  ACTIVE = 'active',
  REVOKED = 'revoked',
}

export enum SustainabilityBadge {
  GREEN_STARTER = 'green_starter',
  SOLAR_CHAMPION = 'solar_champion',
  CLEAN_ENERGY_ADVOCATE = 'clean_energy_advocate',
  CARBON_REDUCER = 'carbon_reducer',
  SUSTAINABILITY_LEADER = 'sustainability_leader',
}

export enum CertificateMonthStatus {
  DOWNLOADABLE = 'downloadable',
  PENDING_CERTIFICATE = 'pending_certificate',
}

export enum CertificateSortOrder {
  NEWEST = 'newest',
  OLDEST = 'oldest',
}